pyproject.toml
[project]
name = "multi-test-tool"
version = "0.1.0"
description = "Multi-purpose testing tool with Kafka producer/consumer capabilities"
authors = [
    {name = "Sanif SS", email = "sanifss@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "kafka-python>=2.0.2",
    "python-dotenv>=1.0.0",
    "flask>=3.0.0",
    "flask-socketio>=5.0.0",
    "python-engineio>=4.0.0",
    "python-socketio>=5.0.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
multi-test-tool = "multi_test_tool.main:main"
multi-test-tool-web = "multi_test_tool.web_app:main"
multi-test-tool-dev = "multi_test_tool.dev_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/multi_test_tool"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88
